feat(plugin): adapter plugin -- druid and duckdb
