+++ core/datacap-ui/public/static/images/plugin/duckdb.png
