--- plugin/datacap-jdbc-druid/src/main/java/io/edurt/datacap/plugin/jdbc/druid/DruidAdapter.java
@@ -1,13 +1,0 @@
-package io.edurt.datacap.plugin.jdbc.druid;
-
-import io.edurt.datacap.spi.adapter.JdbcAdapter;
-import io.edurt.datacap.spi.connection.JdbcConnection;
-
-public class DruidAdapter
-        extends JdbcAdapter
-{
-    public DruidAdapter(JdbcConnection jdbcConnection)
-    {
-        super(jdbcConnection);
-    }
-}
