--- plugin/datacap-jdbc-duckdb/pom.xml
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>io.edurt.datacap</groupId>
-        <artifactId>datacap</artifactId>
-        <version>2024.4.0-SNAPSHOT</version>
-        <relativePath>../../pom.xml</relativePath>
-    </parent>
-
-    <artifactId>datacap-jdbc-duckdb</artifactId>
-    <description>DataCap - DuckDB</description>
-
-    <properties>
-        <duckdb.version>1.0.0</duckdb.version>
-        <plugin.name>jdbc-duckdb</plugin.name>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>io.edurt.datacap</groupId>
-            <artifactId>datacap-spi</artifactId>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>commons-beanutils</groupId>
-            <artifactId>commons-beanutils</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.duckdb</groupId>
-            <artifactId>duckdb_jdbc</artifactId>
-            <version>${duckdb.version}</version>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <artifactId>maven-assembly-plugin</artifactId>
-                <version>${assembly-plugin.version}</version>
-                <configuration>
-                    <finalName>${project.artifactId}</finalName>
-                    <descriptors>
-                        <descriptor>../../configure/assembly/plugin.xml</descriptor>
-                    </descriptors>
-                    <outputDirectory>../../dist/plugins</outputDirectory>
-                </configuration>
-                <executions>
-                    <execution>
-                        <id>make-assembly</id>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>single</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
-</project>
