--- plugin/datacap-jdbc-duckdb/src/main/java/io/edurt/datacap/plugin/jdbc/duckdb/DuckDBAdapter.java
@@ -1,13 +1,0 @@
-package io.edurt.datacap.plugin.jdbc.duckdb;
-
-import io.edurt.datacap.spi.adapter.JdbcAdapter;
-import io.edurt.datacap.spi.connection.JdbcConnection;
-
-public class DuckDBAdapter
-        extends JdbcAdapter
-{
-    public DuckDBAdapter(JdbcConnection jdbcConnection)
-    {
-        super(jdbcConnection);
-    }
-}
