+++ plugin/datacap-plugin-druid/src/main/java/io/edurt/datacap/plugin/DruidPlugin.java
@@ -1,0 +1,5 @@
+package io.edurt.datacap.plugin;
+
+public class DruidPlugin
+        extends Plugin
+{}
