+++ plugin/datacap-plugin-duckdb/src/main/java/io/edurt/datacap/plugin/DuckDBPlugin.java
@@ -1,0 +1,5 @@
+package io.edurt.datacap.plugin;
+
+public class DuckDBPlugin
+        extends Plugin
+{}
