+++ plugin/datacap-plugin-duckdb/src/main/java/io/edurt/datacap/plugin/DuckDBService.java
@@ -1,0 +1,23 @@
+package io.edurt.datacap.plugin;
+
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import io.edurt.datacap.spi.PluginService;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@SuppressFBWarnings(value = {"EI_EXPOSE_REP"})
+public class DuckDBService
+        implements PluginService
+{
+    @Override
+    public String driver()
+    {
+        return "org.duckdb.DuckDBDriver";
+    }
+
+    @Override
+    public String connectType()
+    {
+        return "duckdb";
+    }
+}
